profile

+++ controllers/contactController.js
@@ -3,14 +3,14 @@ const Contact = require('../models/ContactModel');
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access Public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({user_id: req.user.id}).populate("user_id", "name email");
     res.status(200).json(contacts);
 });
 //@desc Create a contact
 //@route POST /api/contacts
-//@access Public
+//@access private
 const createContact = asyncHandler( async (req, res) => {
     const {name, email, phone} = req.body;
     if (!name || !email || !phone) {
@@ -18,14 +18,14 @@ const createContact = asyncHandler( async (req, res) => {
         throw new Error("Please fill all the fields");
     }
     const contact = new Contact({
-        name, email, phone
+        name, email, phone, user_id: req.user.id
     });
     const createdContact = await contact.save();
     res.status(201).json(createdContact);
 });
 //@desc Get a contact
 //@route GET /api/contacts/:id
-//@access Public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -36,7 +36,7 @@ const getContact = asyncHandler(async (req, res) => {
 })
 //@desc Update a contact
 //@route PUT /api/contacts/:id
-//@access Public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const {name, email, phone} = req.body;
     const contact = await Contact.findById(req.params.id);
@@ -44,17 +44,25 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error("Not authorized");
+    }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, {name, email, phone}, {new: true});
     res.status(200).json(updatedContact);
 })
 //@desc Delete a contact
 //@route DELETE /api/contacts/:id
-//@access Public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error("Not authorized");
     }
     await contact.deleteOne();
     res.status(200).json({message: "Contact removed"});
